Replaced `auto` with correct type.
As suggested by Jeroen Bakker.

+++ source/blender/blenkernel/BKE_subdiv_ccg.hh
@@ -291,7 +291,7 @@ void BKE_subdiv_ccg_neighbor_coords_get(const SubdivCCG *subdiv_ccg,
                                         bool include_duplicates,
                                         SubdivCCGNeighbors *r_neighbors);
 
-long BKE_subdiv_ccg_grid_to_face_index(const SubdivCCG *subdiv_ccg, int grid_index);
+int BKE_subdiv_ccg_grid_to_face_index(const SubdivCCG *subdiv_ccg, int grid_index);
 void BKE_subdiv_ccg_eval_limit_point(const SubdivCCG *subdiv_ccg,
                                      const SubdivCCGCoord *coord,
                                      float r_point[3]);
